+++ src/lib/ericchase/Platform/Node/HTML Processor/Example -- Template Processor/main.example.ts
@@ -1,4 +1,5 @@
+import { Path } from '../../Path.js';
 import { LoadIncludeFile, ProcessTemplateFile } from '../TemplateProcessor.js';
 
-await LoadIncludeFile('button', './component/button.html');
-await ProcessTemplateFile('./index.template.html', './index.html');
+await LoadIncludeFile('button', new Path('./component/button.html'));
+await ProcessTemplateFile(new Path('./index.template.html'), new Path('./index.html'));